Code reorg: create a psa module
