+++ Cargo.toml
@@ -25,8 +25,6 @@ console = "0.14"
 dialoguer = "0.8"
 indicatif = "0.17.0-beta.1"
 
-sysinfo = "0.20"
-
 regex = "1"
 
 tar = "0.4.20"
